refactor: 아이콘 위치 및 크기 조정
